requirements.txt
python-telegram-bot>=20.7
ccxt>=4.2.0
pandas>=2.1.0
numpy>=1.24.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0
aiohttp>=3.9.0
